[WOM-165] Set all filter constants of device after powered on or reset
Added LowPassFilterCutoff property to MccsPasdBus device with default
value of 10Hz. This value is used to set all the sensors' low-pass
filter constants of a FNDH/smartbox device after it is powered on.
Reverted simulationMode attribute to SKABaseDevice's implementation in
order to set it to True for functional and integration tests. The
attribute it used to disable the setting of filter constants as it sets
the sensor values of the simulators to values which causes alaram
states.