[WOM-184] Refactor functional and smartbox integration tests
Since the simulator was fixed with WOM-167, the smartbox
device integration tests where antennas are turned on with PowerOnPort()
command is simplified. The initial condition for the tests is now an
initialised FNDH with a smartbox that is turned off.

For all the functional tests, the PaSD simulator is now only
instantiated with no ports turned on - so only a FNDH simulator in
UNINITIALISED status. Therefore the monitor and control scenarios' given
steps are updated to initialise the FNDH and turn on smartbox 1 on FNDH
port 1.