MCCS-2216: Support nulls in values schema
Currently in the Low ITF, we have a unusual station layout, with two "stations" using the same PaSD bus. The MCCS devices can handle this situation if configured correctly. The MCCS helmfile templates don't support this use case, but we can override at helm values level to tweak the configuration.

To do this, we need to remove the two pasdbus devices that the helmfile templates deploy. That can be done by nulling them out in the values file. Helm charts support this – a null is interpreted as unsetting the key. Unfortunately there is a bug in helm (https://github.com/helm/helm/issues/12819) – values files are validated against the json schema before the key is unset. The workaround for this is a bit sad: loosen your schema to explicitly allow null values.